+++ payload/compositor/lumaglass/widgets/UnknownWidget.qml
@@ -10,6 +10,7 @@ Item {
     property bool isLight: false
     property var settings: ({})
     property bool focused: false
+    property bool inner: false
 
     function activate() { /* nothing to do */ }